requirements.txt
greenlet==3.0.3
mysql-connector-python==8.2.0
protobuf==4.21.12
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
